Linux

+++ src/ral/crypt/sha2.cc
@@ -35,8 +35,8 @@
 #include "pal.h"
 
 #define SHFR(x, n) (x >> n)
-#define ROTR(x, n) ((x >> n) | (x << ((sizeof(x) << 3) - n)))
-#define ROTL(x, n) ((x << n) | (x >> ((sizeof(x) << 3) - n)))
+#define ROTR(x, n) ((x >> n) | (x << (static_cast<unsigned>((sizeof(x) << 3) - n))))
+#define ROTL(x, n) ((x << n) | (x >> (static_cast<unsigned>((sizeof(x) << 3) - n))))
 #define CH(x, y, z) ((x & y) ^ (~x & z))
 #define MAJ(x, y, z) ((x & y) ^ (x & z) ^ (y & z))
